Preparing for PHP 8.3 (#311)
* Preparing for PHP 8.3

* some more packages ready, updating phpstan baselines to remove php 7

* tests pass locally

* fixing phpstan, upgrading php-cs-fixer

* versions

+++ library/Zend/Barcode/Renderer/Image.php
@@ -463,9 +463,9 @@ class Zend_Barcode_Renderer_Image extends Zend_Barcode_Renderer_RendererAbstract
                 $this->_resource,
                 $size,
                 $orientation,
-                $position[0] - ($width * cos(pi() * $orientation / 180)),
-                $position[1] + ($width * sin(pi() * $orientation / 180)),
-                $allocatedColor,
+                (int) ($position[0] - ($width * cos(pi() * $orientation / 180))),
+                (int) ($position[1] + ($width * sin(pi() * $orientation / 180))),
+                (int) $allocatedColor,
                 $font,
                 $text
             );
